Convert fork tests to snapshots (#2304)
The previous fork tests were confusing especially with configure()
and were too far displaced from reality with the DummyWriter usage.
