+++ tests/tracer/test_tracer.py
@@ -31,7 +31,6 @@ from ddtrace.tracer import Tracer
 from ddtrace.tracer import _has_aws_lambda_agent_extension
 from ddtrace.tracer import _in_aws_lambda
 from tests.subprocesstest import run_in_subprocess
-from tests.utils import DummyWriter
 from tests.utils import TracerTestCase
 from tests.utils import override_global_config
 
@@ -689,82 +688,6 @@ def test_tracer_fork():
     # Assert the trace got written into the correct queue
     assert len(original_writer._buffer) == 1
     assert len(t.writer._buffer) == 1
-
-
-def test_tracer_trace_across_fork():
-    """
-    When a trace is started in a parent process and a child process is spawned
-        The trace should be continued in the child process
-    """
-    tracer = Tracer()
-    tracer.writer = DummyWriter()
-
-    def task(tracer, q):
-        tracer.writer = DummyWriter()
-        with tracer.trace("child"):
-            pass
-        spans = tracer.writer.pop()
-        q.put([dict(trace_id=s.trace_id, parent_id=s.parent_id) for s in spans])
-
-    # Assert tracer in a new process correctly recreates the writer
-    q = multiprocessing.Queue()
-    with tracer.trace("parent") as parent:
-        p = multiprocessing.Process(target=task, args=(tracer, q))
-        p.start()
-        p.join()
-
-    children = q.get()
-    assert len(children) == 1
-    (child,) = children
-    assert parent.trace_id == child["trace_id"]
-    assert child["parent_id"] == parent.span_id
-
-
-def test_tracer_trace_across_multiple_forks():
-    """
-    When a trace is started and crosses multiple process boundaries
-        The trace should be continued in the child processes
-    """
-    tracer = ddtrace.Tracer()
-    tracer.writer = DummyWriter()
-
-    # Start a span in this process then start a child process which itself
-    # starts a span and spawns another child process which starts a span.
-    def task(tracer, q):
-        tracer.writer = DummyWriter()
-
-        def task2(tracer, q):
-            tracer.writer = DummyWriter()
-
-            with tracer.trace("child2"):
-                pass
-
-            spans = tracer.writer.pop()
-            q.put([dict(trace_id=s.trace_id, parent_id=s.parent_id) for s in spans])
-
-        with tracer.trace("child1"):
-            q2 = multiprocessing.Queue()
-            p = multiprocessing.Process(target=task2, args=(tracer, q2))
-            p.start()
-            p.join()
-
-        task2_spans = q2.get()
-        spans = tracer.writer.pop()
-        q.put([dict(trace_id=s.trace_id, parent_id=s.parent_id, span_id=s.span_id) for s in spans] + task2_spans)
-
-    # Assert tracer in a new process correctly recreates the writer
-    q = multiprocessing.Queue()
-    with tracer.trace("parent") as parent:
-        p = multiprocessing.Process(target=task, args=(tracer, q))
-        p.start()
-        p.join()
-
-    children = q.get()
-    assert len(children) == 2
-    child1, child2 = children
-    assert parent.trace_id == child1["trace_id"] == child2["trace_id"]
-    assert child1["parent_id"] == parent.span_id
-    assert child2["parent_id"] == child1["span_id"]
 
 
 def test_tracer_with_version():
